refactor: bring modules one layer out again
