--- src/world/mod.rs
@@ -1,4 +1,0 @@
-pub mod block;
-pub mod chunks;
-pub mod player;
-pub mod world;
